% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{beautify_Reactome_table}
\alias{beautify_Reactome_table}
\title{Beautifying an enrichment table with Reactome identifiers in it}
\usage{
beautify_Reactome_table(enrich_tbl, Reactome_id_column = "ID")
}
\arguments{
\item{enrich_tbl}{A data frame tabular representation of the enrichment results}

\item{Reactome_id_column}{Character string, specifying which column contains the
Reactome identifier}
}
\value{
A data.frame, ready to be passed on to
\code{DT::datatable(..., escape = FALSE)}, to render the buttons correctly
}
\description{
Beautifying an enrichment table with Reactome identifiers in it
}
\examples{
## TODO
}
