% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biummisc.R
\docType{package}
\name{BIUMmisc}
\alias{BIUMmisc}
\alias{BIUMmisc-package}
\title{BIUMmisc}
\description{
This is a collection of functions and scripts commonly used by the
BIUM-MZ Core Facility
}
\author{
Federico Marini
}
