% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestampr.R
\name{count_up_till_death}
\alias{count_up_till_death}
\title{Prints out the timestamp}
\usage{
count_up_till_death(
  file_out,
  extrainfo = NULL,
  secs_interval = 60,
  max_lines = 10000,
  verbose = TRUE
)
}
\arguments{
\item{file_out}{Character, path to the filename to print out the series of
timestamps}

\item{extrainfo}{Character, specifies the extra information to print out
before the timestamp is recorded. Defaults to}

\item{secs_interval}{Numeric value, specifying the interval of seconds after
which to repeat the timestamp operation}

\item{max_lines}{Integer, number of max lines to record - the time interval
covered will depend on this \emph{and} on the \code{secs_interval}. This is mainly to
avoid having a really endless loop and not being able to stop the process.}

\item{verbose}{Logical, defaults to TRUE}
}
\value{
Invisible NULL - the output is very much the side effect of printing
the timestamp to file
}
\description{
Prints out the timestamp to console and to a file. Useful for diagnosing
potential failures of a terminal server session.
}
\examples{
# count_up_till_death("mytestfile.txt", extrainfo = "YourName")
}
