% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_for_bioinfo.R
\name{create_bioinfo_html}
\alias{create_bioinfo_html}
\title{Create a Quarto html report, Bioinfo-group styled}
\usage{
create_bioinfo_html(
  file_name = NULL,
  report_folder = NULL,
  ext_name = "bioinfo-html",
  expose_aux_files = TRUE
)
}
\arguments{
\item{file_name}{Character string, the name of the file for the report -
no need to specify the qmd extension.}

\item{report_folder}{Character string, pointing to the folder where the
report and its associated files should be created. Defaults to NULL, which
falls back to the current working directory.
A sensible value for this variable could the "vignettes", if used in the
context of a package in development}

\item{ext_name}{Character string, defaulting to "bioinfo-html". Other values
(possibly in other functions) would be "bioinfo-revealjs" or similar...}

\item{expose_aux_files}{Logical value, whether to expose&copy the auxiliary
files (scss and bib) into the toplevel of the specified folder.}
}
\value{
Invisibly, the path to the created quarto source file for the report
}
\description{
Create a Quarto html report, Bioinfo-group styled
}
\examples{
temp_folder <- tempdir()
create_bioinfo_html(file_name = "myreport",
                    report_folder = temp_folder)
}
