% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{create_DEresults}
\alias{create_DEresults}
\title{Title}
\usage{
create_DEresults(
  resuSet = NULL,
  dds_obj,
  contrast_name,
  FDR = 0.05,
  lfc_threshold = 0,
  extended_anno_df,
  species = "Homo_sapiens",
  name_res_entry = contrast_name,
  verbose = TRUE
)
}
\arguments{
\item{resuSet}{A list, provided in the format expected by the resuSet
specifications.
This can be an empty list, if the object was never created beforehand,
otherwise this function can simply add all the set of the results and output
into a new element of an existing list.
Defaults to \code{NULL}, which would lead to the creation of a new list object. If
this is not the desired behavior, please provide an existing object
structured as a \code{resuSet}.}

\item{dds_obj}{A DESeqDataset object, meaningfully related/matched to the
resuSet object.}

\item{contrast_name}{Character string, specifying the name of the contrast to be
extracted}

\item{FDR}{Numeric value, specifying the false discovery rate value to be
used}

\item{lfc_threshold}{Numeric value, corresponding to the logFC theshold to be
used to extract the results - it is relevant in the call to \code{results()} and
it does differ (a lot, conceptually) from the logFC filtering that commonly
is performed post-hoc, after testing against the default null hypothesis
(logFC = 0), which might be too liberal in some cases.}

\item{extended_anno_df}{An annotation data.frame with the following columns:
\code{gene_id} and \code{gene_name} (mandatory), and \code{description} (optional, can be
retrieved via biomaRt)}

\item{species}{Character string, specifying the species for the dataset under
investigation. Defaults to "Homo_sapiens", could also be "Mus_musculus" and
other similarly assembled character strings. This is relevant only when
creating the buttons for the interactive versions of the results tables.}

\item{name_res_entry}{Character string}

\item{verbose}{Logical}
}
\value{
A \code{resuSet} list object.
}
\description{
Title
}
\examples{
# TODO
}
