% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Create folder structure}
\usage{
create_folder_structure(rootname = paste0(Sys.Date(), "_newproject_RENAME_ME"))
}
\arguments{
\item{rootname}{Character string, corresponding to the name of the main folder
to be created. Should ideally follow this schema:
\link{person_identifier}\emph{\link{cooperation_partner_name}}\link{project_related_keyword}}
}
\value{
Creates the designed folder structure in the desired location
}
\description{
Create folder structure
}
\examples{
# An example
}
