% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{create_link_reactome}
\alias{create_link_reactome}
\title{Create a link to Reactome}
\usage{
create_link_reactome(val)
}
\arguments{
\item{val}{Character string, specifying the identifier as in the Reactome
database}
}
\value{
A string with the html code to create a button linking to the
specified identifier
}
\description{
Create a link to Reactome
}
\examples{
# TODO: possibly this could and should live within mosdef...

# https://reactome.org/content/detail/R-HSA-69239
create_link_reactome("R-HSA-69239")
}
