% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{dds_gencode_to_ensembl}
\alias{dds_gencode_to_ensembl}
\title{Gencode to ENSEMBL}
\usage{
dds_gencode_to_ensembl(dds)
}
\arguments{
\item{dds}{A DESeqDataSet object}
}
\value{
A DESeqDataSet object, with updated row names
}
\description{
Gencode to ENSEMBL
}
\examples{
# TODO
}
