% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{deseq_normcounts_with_info}
\alias{deseq_normcounts_with_info}
\title{Normalized counts table, with extra info}
\usage{
deseq_normcounts_with_info(dds, extended_anno_df)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{extended_anno_df}{An annotation data.frame with the following columns:
\code{gene_id} and \code{gene_name} (mandatory), and \code{description} (optional, can be
retrieved via biomaRt)}
}
\value{
A data.frame object, with the normalized counts and some additional
info on the features at hand
}
\description{
Normalized counts table, with extra info
}
\examples{
# TODO
# ext_anno <- merge(anno_df, anns, by.x = "gene_id", by.y = "ensembl_gene_id")

# ...
# writexl::write_xlsx(obj, "exceltable.xlsx")
}
