% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{deseq_tpm_with_info}
\alias{deseq_tpm_with_info}
\title{TPM table with extended information}
\usage{
deseq_tpm_with_info(dds, extended_anno_df)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{extended_anno_df}{An annotation data.frame with the following columns:
\code{gene_id} and \code{gene_name} (mandatory), and \code{description} (optional, can be
retrieved via biomaRt)}
}
\value{
A data.frame object, with the TPM values and some additional
info on the features at hand
}
\description{
TPM table with extended information
}
\examples{
# TODO
}
