% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{exportr}
\alias{exportr}
\title{Export the components of a resuSet object}
\usage{
exportr(resuSet, out_file_prefix = "out_SET", out_folder = "_output")
}
\arguments{
\item{resuSet}{A list, provided in the format expected by the resuSet
specifications}

\item{out_file_prefix}{String character, to be used to prepend all exported files.
Defaults to "out_SET", but can be meaningfully chosen to match the set of
results it refers to.}

\item{out_folder}{String character, specifying in which folder such tabular
outputs will be generated. Defaults to "_output", but can also be called "."
to specify the current directory.}
}
\value{
Invisible NULL if no errors are encountered
}
\description{
Export the components of a resuSet object
}
\examples{
# TODO
}
