% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{fortify_annotations}
\alias{fortify_annotations}
\title{Fortify different set of annotations}
\usage{
fortify_annotations(
  anno_df_1,
  anno_df_2,
  id_1 = "gene_id",
  id_2 = "ensembl_gene_id",
  gene_name_1 = "gene_name",
  gene_name_2 = "external_gene_name",
  dds = NULL,
  verbose = TRUE
)
}
\arguments{
\item{anno_df_1}{A data.frame object, corresponding to the first annotation
table}

\item{anno_df_2}{A data.frame object, corresponding to the second annotation
table}

\item{id_1}{Character string, specifying which column in the first table is
to be used as identifier (defaults to \code{gene_id})}

\item{id_2}{Character string, specifying which column in the second table is
to be used as identifier (defaults to \code{ensembl_gene_id}, for biomaRt-like
objects)}

\item{gene_name_1}{Character string, specifying which column in the first
table is to be used as gene name (defaults to \code{gene_name})}

\item{gene_name_2}{Character string, specifying which column in the second
table is to be used as identifier (defaults to \code{external_gene_name}, again
as could be taken out e.g. from biomaRt)}

\item{dds}{A DESeqDataset object}

\item{verbose}{Logical, controlling the verbosity level of the function}
}
\value{
A list, including the fortified annotation data frame (in the
\code{anno_df} element), and (if provided as input) a dds object (in the \code{dds}
object) with the information stored in the \code{rowData} slot.
}
\description{
Fortify different set of annotations
}
\examples{
# TODO
}
