% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{mamma_singlecell}
\alias{mamma_singlecell}
\alias{matmetmaker_RNAseqsinglecell}
\title{Title TODO}
\usage{
mamma_singlecell(
  sce_object,
  template_text = NULL,
  packages = NULL,
  FDRthreshold = NULL,
  de_framework = "muscat",
  add_citations = TRUE
)

matmetmaker_RNAseqsinglecell(
  sce_object,
  template_text = NULL,
  packages = NULL,
  FDRthreshold = NULL,
  de_framework = "muscat",
  add_citations = TRUE
)
}
\arguments{
\item{sce_object}{TODO}

\item{template_text}{TODO}

\item{packages}{TODO}

\item{FDRthreshold}{TODO}

\item{de_framework}{TODO}

\item{add_citations}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
