% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{plot_totcounts}
\alias{plot_totcounts}
\title{Plot total number of counts}
\usage{
plot_totcounts(dds, group)
}
\arguments{
\item{dds}{A DESeqDataset object}

\item{group}{Character string, indicating one of the names of the columns in
the \code{colData} slot of \code{dds}}
}
\value{
A ggplot object
}
\description{
Plot total number of counts
}
\examples{
# TODO
}
