% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_funcs.R
\name{resuset_to_gtl}
\alias{resuset_to_gtl}
\title{Convert a resuSet into a GeneTonicList}
\usage{
resuset_to_gtl(resuSet, result_name, which_enrich = "topGO_tbl", dds, anno_df)
}
\arguments{
\item{resuSet}{A list, provided in the format expected by the resuSet
specifications}

\item{result_name}{A character, specifying for which result name in the resuSet
object the conversion should be performed}

\item{which_enrich}{Character string, specifying which enrichment results to
use. Currently defaults to "topGO_tbl", will be extended to other formats -
TODO}

\item{dds}{A DESeqDataset object, meaningfully related/matched to the
resuSet object.}

\item{anno_df}{A data.frame as specified in the requirements expected by
GeneTonic, i.e. where at least the columns \code{gene_id} and \code{gene_name} are
specified}
}
\value{
A GeneTonicList object, ready to be provided to the GeneTonic
functions
}
\description{
Convert a resuSet into a GeneTonicList object
}
\examples{
# TODO

}
