% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesheet_management.R
\name{samplesheet_creator}
\alias{samplesheet_creator}
\title{Title}
\usage{
samplesheet_creator(
  n_samples = 4,
  ss_filename,
  ss_dir = ".",
  extra_fields = NULL
)
}
\arguments{
\item{n_samples}{Numeric value, corresponds to the number of samples included}

\item{ss_filename}{Character string, basename of the excel file that will
get generated}

\item{ss_dir}{Character string, folder where the sample sheet will be
generated into. Defaults to "." (the current working directory).}

\item{extra_fields}{Vector of character strings, specifying the names of the
columns for extra fields to be added.}
}
\value{
The character string corresponding to the path to the Excel sample
sheet created. The Excel sheet is created as a side effect.
}
\description{
Title
}
\examples{
mysheet <- samplesheet_creator(ss_filename = "mysheet",
                               ss_dir = tempdir())
mysheet
}
