% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIUMmisc-pkg.R
\docType{package}
\name{BIUMmisc-pkg}
\alias{BIUMmisc}
\alias{BIUMmisc-package}
\alias{BIUMmisc-pkg}
\title{BIUMmisc}
\description{
This is a collection of functions and scripts commonly used by the
BIUM-MZ Core Facility and the Bioinformatics group at the IMBEI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imbeimainz/BIUMmisc}
  \item Report bugs at \url{https://github.com/imbeimainz/BIUMmisc/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
  \item a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
  \item a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
}

}
\keyword{internal}
