% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessioninfo_extra.R
\name{sessioninfo_extra}
\alias{sessioninfo_extra}
\title{Title}
\usage{
sessioninfo_extra(to_file = TRUE, out_file = NULL)
}
\arguments{
\item{to_file}{todo}

\item{out_file}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
# todo
}
